Refactored channel model to its own app
* Created application channels
* Moved all models, views and helpers to the new app and removed from
the talks one
* Clean up migrations from the talk app
