+++ web/talks/api.py
@@ -6,18 +6,11 @@ from rest_framework.response import Response
 
 from django.db.models.aggregates import Count
 
-from .models import Channel
 from .models import Playlist
 from .models import Talk
 
-from .serializers import ChannelSerializer
 from .serializers import PlaylistSerializer
 from .serializers import TalkSerializer
-
-
-class ChannelViewSet(viewsets.ModelViewSet):
-    queryset = Channel.objects.all()
-    serializer_class = ChannelSerializer
 
 
 class PlaylistViewSet(viewsets.ModelViewSet):
