+++ web/talks/urls.py
@@ -18,13 +18,11 @@ from .views import RSSLatestView
 
 from rest_framework import routers
 
-from .api import ChannelViewSet
 from .api import PlaylistViewSet
 from .api import TalkViewSet
 from .api import RandomTalkView
 
 router = routers.DefaultRouter()
-router.register(r'channel', ChannelViewSet)
 router.register(r'playlist', PlaylistViewSet)
 router.register(r'talk', TalkViewSet)
